+++ src/components/smallComps/badge.jsx
@@ -2,35 +2,30 @@ import React, { Component } from "react";
 
 class Badge extends Component {
   render() {
-    return (
-      <li className="pb-5 flex flex-col items-center">
-        <div
-          className={
-            "bg-white shadow-md rounded-full flex justify-center items-center" +
-            " w-" +
-            this.props.size +
-            " h-" +
-            this.props.size
-          }
-        >
-          <img
-            className={"p-" + this.props.imgPadding}
-            src={this.props.img}
-            alt={this.props.alt}
-          />
-        </div>
-        <div className="pt-2 w-32 overflow-hidden text-center text-lg font-black text-inset">
-          {this.props.text}
-        </div>
-      </li>
-    );
+    if (this.props.size === "sm") {
+      return (
+        <li className="pb-5 flex flex-col items-center">
+          <div className="bg-white shadow-md rounded-full flex justify-center items-center w-20 h-20">
+            <img className="p-2" src={this.props.img} alt={this.props.alt} />
+          </div>
+          <div className="pt-2 w-32 overflow-hidden text-center text-lg font-black text-inset">
+            {this.props.text}
+          </div>
+        </li>
+      );
+    } else if (this.props.size === "lg") {
+      return (
+        <li className="pb-5 flex flex-col items-center">
+          <div className="bg-white shadow-md rounded-full flex justify-center items-center w-28 h-28">
+            <img className="p-3" src={this.props.img} alt={this.props.alt} />
+          </div>
+          <div className="pt-2 w-32 overflow-hidden text-center text-lg font-black text-inset">
+            {this.props.text}
+          </div>
+        </li>
+      );
+    }
   }
 }
-
-Badge.defaultProps = {
-  size: "28",
-  text: "Add Text Here",
-  imgPadding: "3",
-};
 
 export default Badge;
